project/auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
sqlalchemy = "^2.0.32"
uvloop = "^0.19.0"
dependency-injector = "4.42.0b1"
pydantic-settings = "^2.4.0"
passlib = "^1.7.4"
bcrypt = "^4.2.0"
email-validator = "^2.2.0"
pyjwt = "^2.9.0"
python-multipart = "^0.0.9"
uvicorn = "^0.30.6"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
aiokafka = "^0.11.0"
python-schema-registry-client = "^2.6.0"
dataclasses-avroschema = "^0.62.7"
kafka-python = "^2.0.2"
pytz = "^2024.1"
orjson = "^3.10.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
